+++ Editor/Source/EditorLayer.cpp
@@ -1,0 +1,12 @@
+#include "EditorLayer.h"
+
+namespace Core
+{
+    void EditorLayer::OnAttach()
+    {
+    }
+
+    void EditorLayer::OnUpdate()
+    {
+    }
+}