+++ Editor/Source/EditorLayer.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "Core.h"
+
+namespace Core
+{
+    class EditorLayer : public Layer
+    {
+    public:
+        EditorLayer() {};
+        ~EditorLayer() {};
+
+        void OnAttach();
+        void OnUpdate();
+    };
+}