+++ Engine/Source/Core.h
@@ -6,3 +6,8 @@
 #include "Core/CeAssert.h"
 #include "Core/Engine.h"
 #include "Core/Logger.h"
+#include "Core/Window.h"
+#include "Core/Input.h"
+
+#include "Core/Layer/Layer.h"
+#include "Core/Layer/LayerStack.h"