+++ Engine/Source/Core/Engine.cpp
@@ -1,51 +1,76 @@
 #include "Engine.h"
 #include "Logger.h"
+#include "Input.h"
+
+#include "Layer/LayerStack.h"
 
 namespace Core
 {
-    static Application *GApp = nullptr;
+    static Engine::State state;
 
     void Engine::PreInit()
     {
         Logger::Settings LogSettings; // TODO: Maybe parse a file?
         Logger::Init(LogSettings);
+        Input::Init();
+        LayerStack::Init();
+
+        Window::Information WindowInformation;
+        WindowInformation.X = WindowInformation.Y = 100;
+        WindowInformation.Width = 1280;
+        WindowInformation.Height = 720;
+        WindowInformation.Title = "Hello Core";
+        state._Window = new Window(WindowInformation);
     }
 
     void Engine::Init()
     {
-        if (GApp)
-            GApp->Init();
+        if (state._GApp)
+            state._GApp->Init();
     }
 
     void Engine::Update()
     {
-        if (GApp)
-            GApp->Update();
+        LayerStack::Update();
+        state._Window->Update();
+
+        if (state._GApp)
+            state._GApp->Update();
     }
 
     void Engine::Render()
     {
-        if (GApp)
-            GApp->Render();
+        if (state._GApp)
+            state._GApp->Render();
     }
 
     void Engine::Shutdown()
     {
-        if (GApp)
+        if (state._GApp)
         {
-            GApp->Shutdown();
-            delete GApp;
-            GApp = nullptr;
+            state._GApp->Shutdown();
+            delete state._GApp;
+            state._GApp = nullptr;
         }
+
+        delete state._Window;
+        Logger::Shutdown();
+        LayerStack::Shutdown();
+        Input::Shutdown();
     }
 
     bool Engine::ShouldRun()
     {
-        return true;
+        return state._Window->ShouldRun();
+    }
+
+    Window *Engine::GetWindow()
+    {
+        return state._Window;
     }
 
     void Engine::FeedApplication(Application *_App)
     {
-        GApp = _App;
+        state._GApp = _App;
     }
 }