+++ Engine/Source/Core/Layer/Layer.cpp
@@ -1,0 +1,24 @@
+#include "Layer.h"
+
+namespace Core
+{
+    Layer::Layer()
+    {
+    }
+
+    Layer::~Layer()
+    {
+    }
+
+    void Layer::OnAttach()
+    {
+    }
+
+    void Layer::OnUpdate()
+    {
+    }
+
+    void Layer::OnDetach()
+    {
+    }
+}