+++ Engine/Source/Core/Layer/Layer.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "Base.h"
+
+namespace Core
+{
+    class CE_API Layer
+    {
+    public:
+        Layer();
+        ~Layer();
+
+        virtual void OnAttach();
+        virtual void OnUpdate();
+        virtual void OnDetach();
+    };
+}