+++ CHANGELOG.md
@@ -4,3 +4,14 @@
 - Started logger
 - Added application and engine concept
 - Structure
+
+# 0.0.2 - 09.05.2024
+
+- Added GLFW
+- Added Scripts
+- Added Window
+- Engine State structure
+- Started Size Modes
+- Added more configuration to window (decoration, VSync, ...)
+- Started Input
+- Added Layers